rename tree_env to branching_env
